Rename __init__ to test_base and update imports
Pytest guidelines state that the tests dir should not be a module
